glados_riva/scripts/requirements.txt
discord.py==2.3.2
num2words==0.5.12
nvidia_riva_client==2.8.0
